Move hvac to client.py
